Revert "Calculate"
